Refactor: Split `dht` package into multiple smaller packages
This commit splits the `dht` package into smaller, more focused packages to improve code organization and maintainability. The following new packages were created:

- `arena`:
- `logger`:
- `peer`:
- `remoteNode`:
- `routingTable`:

Each of these packages now has a clear responsibility and contains only the code that is necessary to fulfill that responsibility.

This commit does not change the functionality of the code, only its organization.